added contextual dropdown for runtime configuration selection
